+++ test/getmodule/getmodule_getter.ts
@@ -1,7 +1,7 @@
 import Vuex from 'vuex'
 import Vue from 'vue'
 Vue.use(Vuex)
-import { getModule, Module, Action, Mutation, VuexModule } from '..'
+import { getModule, Module, Action, Mutation, VuexModule } from '../..'
 import { expect } from 'chai'
 
 const store = new Vuex.Store({})
